correção tipo variavel v_diaria

+++ Ex015.py
@@ -12,7 +12,7 @@ def f_calcula_diaria_km(v_diaria,v_km):
 # Input float dos dados com try-except.
 try:
     # Recebe float da quantidade de diárias com o carro.
-    v_diaria = float(input("Digite o número de diárias: "))    
+    v_diaria = int(input("Digite o número de diárias: "))    
     # Se o valor da diária for menor que 1, abre except e fecha o programa.
     if v_diaria < 1:
         raise
